Add AFDM
The AFDM is the main object of the argumentation lib: it holds Arguments
and Attacks, represents the argumentation graph, and computes the grounded
extension.

Note: previous versions of the AFDM class used a Graph to hold Arguments
and Attacks, but the AFDM also contained Arguments, which seemed a bit
redundant. Also, the binary relationship R as a list of Attacks was not
optimal for performances, especially when getting attackers or attacked
arguments. This version merges the Graph with the AFDM, to avoid duplicating
functions, and uses MultiSets to optimize the attacks.